Tweek the check function
